server/pyproject.toml
[project]
name = "postgres"
version = "0.1.0"
description = "A simple MCP postgres server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.2.0",
    "psycopg[binary]>=3.1.18"
] 


[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["main"]  
